+++ RELEASE_NOTES.md
@@ -83,6 +83,9 @@ Release Notes
 * Dedicated Hosts: Support for Host Groups and Hosts
 * WebApps: Added support for Node 14, 16 and 18
 
+## 1.7.11
+* Storage Account Queues: Support for adding metadata
+
 ## 1.7.10
 * Container Groups and Container Apps: Support for link_to_identity for ACR managed identities.